+++ challenges/src/main/java/com/cart/Cart.java
@@ -1,0 +1,19 @@
+package com.cart;
+
+import java.util.List;
+
+public class Cart {
+    private List<Product> products;
+    private DiscountService discountService;
+
+    public Cart(List<Product> products, DiscountService discountService) {
+        this.products = products;
+        this.discountService = discountService;
+    }
+
+    public double calculateTotal() {
+        return products.stream()
+            .mapToDouble(p -> discountService.applyDiscount(p))
+            .sum();
+    }
+}
