+++ challenges/src/main/java/com/cart/Main.java
@@ -1,0 +1,7 @@
+package com.cart;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("main desde challenges");
+    }
+}
