--- challenges/week-01-cart-discounts/src/com/walmarttech/cart/Cart.java
@@ -1,17 +1,0 @@
-import java.util.List;
-
-public class Cart {
-    private List<Product> products;
-    private DiscountService discountService;
-
-    public Cart(List<Product> products, DiscountService discountService) {
-        this.products = products;
-        this.discountService = discountService;
-    }
-
-    public double calculateTotal() {
-        return products.stream()
-            .mapToDouble(p -> discountService.applyDiscount(p))
-            .sum();
-    }
-}
