--- challenges/week-01-cart-discounts/src/com/walmarttech/cart/DiscountService.java
@@ -1,8 +1,0 @@
-public class DiscountService {
-    public double applyDiscount(Product product) {
-        if (product.hasDiscount() && product.getQuantity() >= 3) {
-            return product.getTotalPrice() * 0.9; // 10% descuento
-        }
-        return product.getTotalPrice();
-    }
-}
