+++ debugging/src/main/java/com/discount/Main.java
@@ -1,0 +1,7 @@
+package com.discount;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("main discount");
+    }
+}
