--- debugging/week-01-missing-discount/src/com/walmarttech/discount/DiscountCalculator.java
@@ -1,9 +1,0 @@
-public class DiscountCalculator {
-
-    public double calculateDiscount(Customer customer, double totalAmount) {
-        if (customer.getMembershipLevel() == "Gold") {
-            return totalAmount * 0.2; // 20% de descuento
-        }
-        return 0.0;
-    }
-}
