--- debugging/week-01-missing-discount/test/com/walmarttech/discount/DiscountCalculatorTest.java
@@ -1,21 +1,0 @@
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-public class DiscountCalculatorTest {
-
-    @Test
-    public void testGoldCustomerGetsDiscount() {
-        Customer customer = new Customer("Ana", "Gold");
-        DiscountCalculator calculator = new DiscountCalculator();
-        double discount = calculator.calculateDiscount(customer, 100.0);
-        assertEquals(20.0, discount, 0.01); // Espera 20% de descuento
-    }
-
-    @Test
-    public void testNonGoldCustomerGetsNoDiscount() {
-        Customer customer = new Customer("Luis", "Silver");
-        DiscountCalculator calculator = new DiscountCalculator();
-        double discount = calculator.calculateDiscount(customer, 100.0);
-        assertEquals(0.0, discount, 0.01); // Sin descuento
-    }
-}
